app/pyproject.toml
[project]
name = "file-ai-coding-challenge-app"
version = "0.1.0"
description = "Dependencies for running the app in production -- need a separate file for Docker COPY."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "xlsxwriter>=3.2.9",
]

[dependency-groups]
dev = []
